# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive file server API built with Go. It supports file and folder management, sharing, real-time updates via WebSockets, and more. All protected endpoints require a Bearer Token for authorization. The API implements rate limiting: most routes are limited to 60 requests/second per IP, with stricter limits on login endpoints.",
        "title": "File Server API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/nodes/{nodeId}/restore": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Restores a file or folder (and all of its contents) from the trash to its original location. Fails if a node with the same name already exists in the target location, unless 'renameOnConflict' is true.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nodes"
                ],
                "summary": "Restore a node from trash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node ID to restore",
                        "name": "nodeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "If true, renames the node (e.g., 'file (1).txt') on conflict instead of failing.",
                        "name": "renameOnConflict",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serwer-plikow_internal_models.RichNode"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Node not found in trash",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict - a node with the same name already exists in the original location",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/trash": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all files and folders currently in the user's trash.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trash"
                ],
                "summary": "List trash contents",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order. Comma-separated list of fields. Use '-' for descending. E.g., 'type,-name'",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/serwer-plikow_internal_models.RichNode"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/trash/purge": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently deletes all files and folders from the user's trash. This action cannot be undone.",
                "tags": [
                    "trash"
                ],
                "summary": "Purge trash",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/trash/{nodeId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently deletes a single file or folder (and all its contents) from the user's trash. This action cannot be undone.",
                "tags": [
                    "trash"
                ],
                "summary": "Permanently delete a single item from trash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ID of the node to permanently delete from trash",
                        "name": "nodeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Node not found in trash",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "serwer-plikow_internal_models.RichNode": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "_vx2a-43VqRT5wz_s9u4"
                },
                "is_favorited": {
                    "type": "boolean",
                    "example": true
                },
                "is_shared": {
                    "type": "boolean",
                    "example": false
                },
                "mime_type": {
                    "type": "string",
                    "example": "application/pdf"
                },
                "modified_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "Raport_Q3.docx"
                },
                "node_type": {
                    "type": "string",
                    "example": "file"
                },
                "original_parent_id": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/serwer-plikow_internal_models.RichNodeOwner"
                },
                "parent_id": {
                    "type": "string",
                    "example": "fLW5kAh2ia9vYmjMnU4nZ"
                },
                "permissions": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer",
                    "example": 123456
                }
            }
        },
        "serwer-plikow_internal_models.RichNodeOwner": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "Administrator"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}